fix integer type in OH_NativeWindow_NativeWindowHandleOpt

+++ engine/src/flutter/shell/platform/ohos/ohos_external_texture_gl.cpp
@@ -385,7 +385,7 @@ void OHOSExternalTextureGL::ProducePixelMapToNativeImage() {
         << ret;
   }
   
-  int32_t usage = 0;
+  uint64_t usage = 0;
   OH_NativeWindow_NativeWindowHandleOpt(native_window_, GET_USAGE, &usage);
   usage |= NATIVEBUFFER_USAGE_CPU_READ;
   OH_NativeWindow_NativeWindowHandleOpt(native_window_, SET_USAGE, usage);
